moved TEvAuditLog to the header file
